Adding ablation scenario to the new generation structure
