+++ sparkle/configurator/__init__.py
@@ -1,5 +1,4 @@
 """This package provides configurator support for Sparkle."""
-from sparkle.configurator.configurator import Configurator, ConfigurationScenario
-from sparkle.configurator.ablation import AblationScenario
-
+from sparkle.configurator.configurator import (Configurator, ConfigurationScenario,
+                                               AblationScenario)
 from sparkle.configurator import implementations
